backend/pyproject.toml
[tool.poetry]
name = "open-advocacy"
version = "0.1.0"
description = ""
authors = ["Matt Hendrick <75456865+matt-hendrick@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
pydantic = "^2.11.1"
sqlalchemy = "^2.0.40"
alembic = "^1.15.2"
pytest = "^8.3.5"
httpx = "^0.28.1"
python-dotenv = "^1.1.0"
ruff = "^0.11.2"
mypy = "^1.15.0"
asyncpg = "^0.30.0"
aiosqlite = "^0.21.0"
pydantic-settings = "^2.8.1"
aiohttp = "^3.11.16"
shapely = "^2.1.0"
gunicorn = "^23.0.0"
python-jose = {version = "^3.4.0", extras = ["cryptography"]}
python-multipart = "^0.0.20"
bcrypt = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
